requirements.txt
fastapi
uvicorn
tensorflow
numpy
Pillow
psycopg2
python-multipart
